chore: add cli logic to gamespider
